// src/store/login/type.ts
/*
 * @Author: your name
 * @Date: 2021-08-09 06:14:06
 * @LastEditTime: 2021-08-10 00:24:55
 * @LastEditors: Please set LastEditors
 * @Description: Login 功能的 vuex 模块
 * @FilePath: \vue3-ts-cms\src\store\login\type.ts
 */
export interface ILoginState {
  token: string
  userInfo: any
  userMenus: any
}